# api/openapi/medicore-v1.yaml
openapi: 3.1.0
info:
  title: MediCore Hospital API
  description: "Generated contract for the demonstrated MediCore route families: auth,\
    \ patients, appointments, admissions, command-center dashboard, and audit. MediCore\
    \ is an educational Training/Portfolio system operating on synthetic data only:\
    \ it is non-clinical and makes no production, regulatory, or availability claim.\
    \ Every authenticated operation rides the single server-issued bearer token; scope\
    \ and authority are server-owned. Client errors use the shared ApiError body;\
    \ security refusals (401/403) use one generic non-enumerating error body."
  version: v1
servers:
- url: /
  description: Same-origin review deployment
security:
- bearerAuth: []
...
paths:
  /api/patients/{id}:
    get:
      tags:
      - patient-controller
      operationId: getPatient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PatientResponse"
        "401":
          description: "Missing, invalid, or expired bearer token; the acting session\
            \ is gone. The body is one generic non-enumerating error field."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "403":
          description: The authenticated acting context is not permitted to use this
            operation. The body is one generic non-enumerating error field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "400":
          description: A path value is malformed for its typed target (shared ApiError
            body).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No resource referenced by the path exists inside the acting
            scope (shared ApiError body).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      tags:
      - patient-controller
      operationId: updatePatient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePatientRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PatientResponse"
        "401":
          description: "Missing, invalid, or expired bearer token; the acting session\
            \ is gone. The body is one generic non-enumerating error field."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "403":
          description: The authenticated acting context is not permitted to use this
            operation. The body is one generic non-enumerating error field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "400":
          description: "Malformed request body, malformed field value, or a failed\
            \ validation constraint (shared ApiError body)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No resource referenced by the path exists inside the acting
            scope (shared ApiError body).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/patients:
    get:
      tags:
      - patient-controller
      operationId: listPatients
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PatientResponse"
        "401":
          description: "Missing, invalid, or expired bearer token; the acting session\
            \ is gone. The body is one generic non-enumerating error field."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "403":
          description: The authenticated acting context is not permitted to use this
            operation. The body is one generic non-enumerating error field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
    post:
      tags:
      - patient-controller
      operationId: createPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientRequest"
        required: true
      responses:
        "200":
          description: The created patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientResponse"
        "409":
          description: A patient with this medical record number already exists (shared
            ApiError body; no partial write and no overwrite).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: "Missing, invalid, or expired bearer token; the acting session\
            \ is gone. The body is one generic non-enumerating error field."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "403":
          description: The authenticated acting context is not permitted to use this
            operation. The body is one generic non-enumerating error field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "400":
          description: "Malformed request body, malformed field value, or a failed\
            \ validation constraint (shared ApiError body)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/appointments:
    get:
      tags:
      - appointment-controller
      operationId: listAppointments
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentResponse"
        "401":
          description: "Missing, invalid, or expired bearer token; the acting session\
            \ is gone. The body is one generic non-enumerating error field."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "403":
          description: The authenticated acting context is not permitted to use this
            operation. The body is one generic non-enumerating error field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
    post:
      tags:
      - appointment-controller
      operationId: createAppointment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
        required: true
      responses:
        "200":
          description: The created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
        "404":
          description: The referenced patient or professional does not exist inside
            the acting branch (shared ApiError body; cross-branch references answer
            the same 404).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The requested window is not contained in one availability interval
            or overlaps an existing non-cancelled appointment (shared ApiError body;
            no partial write).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: "Missing, invalid, or expired bearer token; the acting session\
            \ is gone. The body is one generic non-enumerating error field."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "403":
          description: The authenticated acting context is not permitted to use this
            operation. The body is one generic non-enumerating error field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "400":
          description: "Malformed request body, malformed field value, or a failed\
            \ validation constraint (shared ApiError body)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/appointments/{id}:
    get:
      tags:
      - appointment-controller
      operationId: getAppointment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
        "401":
          description: "Missing, invalid, or expired bearer token; the acting session\
            \ is gone. The body is one generic non-enumerating error field."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "403":
          description: The authenticated acting context is not permitted to use this
            operation. The body is one generic non-enumerating error field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "400":
          description: A path value is malformed for its typed target (shared ApiError
            body).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No resource referenced by the path exists inside the acting
            scope (shared ApiError body).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      tags:
      - appointment-controller
      operationId: deleteAppointment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
        "401":
          description: "Missing, invalid, or expired bearer token; the acting session\
            \ is gone. The body is one generic non-enumerating error field."
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "403":
          description: The authenticated acting context is not permitted to use this
            operation. The body is one generic non-enumerating error field.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: authentication required
                required:
                - error
        "400":
          description: A path value is malformed for its typed target (shared ApiError
            body).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No resource referenced by the path exists inside the acting
            scope (shared ApiError body).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
    UpdatePatientRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        phone:
          type: string
        email:
          type: string
        address:
          type: string
      required:
      - fullName
    PatientResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        branchId:
          type: string
          format: uuid
        medicalRecordNumber:
          type: string
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
        sex:
          type: string
        phone:
          type: string
        email:
          type: string
        nationalId:
          type: string
        address:
          type: string
        active:
          type: boolean
...
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
...
    CreatePatientRequest:
      type: object
      properties:
        medicalRecordNumber:
          type: string
          minLength: 1
        fullName:
          type: string
          minLength: 1
        dateOfBirth:
          type: string
          format: date
        sex:
          type: string
        phone:
          type: string
        email:
          type: string
        nationalId:
          type: string
        address:
          type: string
      required:
      - fullName
      - medicalRecordNumber
...
    CreateAppointmentRequest:
      type: object
      properties:
        patientId:
          type: string
          format: uuid
        professionalId:
          type: string
          format: uuid
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          format: int32
          maximum: 480
          minimum: 5
        type:
          type: string
          minLength: 1
        status:
          type: string
          minLength: 1
          pattern: scheduled|confirmed|completed|cancelled
      required:
      - durationMinutes
      - patientId
      - professionalId
      - scheduledAt
      - status
      - type
    AppointmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        branchId:
          type: string
          format: uuid
        patientId:
          type: string
        professionalId:
          type: string
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
          format: int32
        endsAt:
          type: string
        type:
          type: string
        status:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: The access token issued by POST /api/auth/login (or a successful
        acting-context switch); it alone selects the acting context.
      scheme: bearer
      bearerFormat: JWT